requirements.txt
python-telegram-bot==20.7
opencv-python
numpy
Pillow
face_recognition
